Add bucket prefix filtering and prefix stripping transformation
This commit adds two generic concepts to buckets:
    - filtering, that allows to not specify which resources subset is
      needed by a task, allowing the qbox to download only that
    - transformation, that allows to make some changes to resources
      after they are downloaded but before being made available to the
      task as resource

There is one implementation of each:
    - PrefixFiltering, which allows to use only a "subdirectory" in a
      bucket
    - the StripPrefix transformation, that allows to stip a prefix from
      the object path when copying it in the task's resources directory

Combined, those two options allow to use a "subdirectory" in a bucket to
be used as a lightweight bucket emulation, in particular useful when
needing many bucket-like (e.g for many tasks). Using many buckets offers
pretty bad performance (high creation cost, poor scaling at least in the
case of Ceph), while using a prefix in a bucket is mostly free (no
creation cost as a prefix does not even need to be declared).

Fix QNET-3316